feat: move Convergence from Studio to Overview
- convergence.{js,css} -> overview-convergence.{js,css}; panel + tab on
  overview.html; satellite reads OV.state / OV.findOverlappingData / the
  utils.js buildXrefBadges instead of window._studio*
- first activation awaits the hub's ensureData() (own baseline/bootstrap
  fetches deleted); Refresh routes through OV.refreshData(); thumbnail
  previews call ../studio/api/thumbnail cross-prefix
- dispatch-to-queue dropped: the artifact/reel queues are in-page Studio
  state a standalone page cannot reach
- GET/PUT /api/convergence/offsets + _clean_convergence_offsets move from
  the studio blueprint to overview.py (no sheet dependency; the moved JS's
  relative calls now resolve on /overview/)
- Studio detached: tab button, panel, css/script tags, syncPreviewTab
  branch, resize hook, checkConvergenceTabVisibility, and the
  _studioFindOverlappingData/_studioBuildXrefBadges/_studioRender*Queue
  exports removed
- offsets API tests move to tests/test_overview.py against /overview/
