+++ assets/web/studio-intake.js
@@ -6,10 +6,10 @@
  * namespace; the hub calls back in via same-named guarded delegators. Loaded by
  * studio.html immediately after studio.js.
  *
- * Kept in the hub (NOT moved here; reached via STUDIO): `buildXrefBadges` (also
- * the legacy `window._studioBuildXrefBadges` that convergence.js reads) and the
- * `ss*Thumb*` lazy-thumbnail cluster (`ssClearPending`), which the hub's queue
- * cards share.
+ * Kept in the hub (NOT moved here; reached via STUDIO): the `ss*Thumb*`
+ * lazy-thumbnail cluster (`ssClearPending`), which the hub's queue cards
+ * share. `buildXrefBadges` is a utils.js ambient global (shared with the
+ * Overview page).
  */
 
 (function () {
@@ -20,7 +20,6 @@
   var attachQueueScrubbers = STUDIO.attachQueueScrubbers,
     buildQueueCardThumb = STUDIO.buildQueueCardThumb,
     buildXrefBadges = STUDIO.buildXrefBadges,
-    checkConvergenceTabVisibility = STUDIO.checkConvergenceTabVisibility,
     findIntakeInQueue = STUDIO.findIntakeInQueue,
     findOverlappingData = STUDIO.findOverlappingData,
     intakeAddItem = STUDIO.intakeAddItem,
@@ -193,7 +192,6 @@
     var threshold = parseInt((qs("#intakeClusterThreshold") || {}).value, 10) || 10;
     state.intakeClusters = clusterIntakeEvents(intakeClusterSource(), threshold);
     renderIntake(hasNew);
-    checkConvergenceTabVisibility();
     refreshMetadataIfActive();
     return true;
   }
@@ -832,7 +830,6 @@
       return false;
     }
     renderTranscriptIntake();
-    checkConvergenceTabVisibility();
     refreshMetadataIfActive();
     return true;
   }
@@ -873,7 +870,6 @@
           }
           state.trIntakeClusters = clusterTranscriptMarks(allItems, threshold);
           renderTranscriptIntake();
-          checkConvergenceTabVisibility();
           refreshMetadataIfActive();
         });
       })
